+++ documentation/source/history.md
@@ -1,8 +1,8 @@
 ---
-title: EastSyriacMarcusNew - Version History
-fontversion: 1.101
+title: East Syriac Marcus New - Version History
+fontversion: 1.105
 ---
 
-### xx xxx 2025 (WSTech team) EastSyriacMarcusNew Version 1.xxx
+### xx xxx 2025 (WSTech team) East Syriac Marcus New Version 1.xxx
 - First version of font