+++ documentation/source/productsite/history.md
@@ -1,4 +1,4 @@
 
-### xx xxx 2025 (WSTech team) EastSyriacMarcusNew Version 1.xxx
+### xx xxx 2025 (WSTech team) East Syriac Marcus New Version 1.xxx
 - First version of font